# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Batuan, Fosil dan Sumber Daya Geologi",
        "title": "Museum Geologi Bandung",
        "contact": {},
        "version": "1.0"
    },
    "host": "https://sbc-sebatcabut.herokuapp.com/",
    "basePath": "/v1",
    "paths": {
...
        "/fosil": {
            "post": {
                "description": "Create a new Fosil with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fosil"
                ],
                "summary": "Create a new Fosil",
                "parameters": [
                    {
                        "description": "The fosil to create",
                        "name": "fosil",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Fosil"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.FosilResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.FosilResponse"
                        }
                    }
                }
            }
        },
        "/fosil/{fosilId}": {
            "get": {
                "description": "Get a Fosil by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fosil"
                ],
                "summary": "Get Fosil by ID",
                "operationId": "get-fosil-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fosil ID",
                        "name": "fosilId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.FosilResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.FosilResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit an existing Fosil",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fosil"
                ],
                "summary": "Edit an existing Fosil",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the Fosil to edit",
                        "name": "fosilId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fosil object that needs to be edited",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Fosil"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.FosilResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.FosilResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a fosil by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fosil"
                ],
                "summary": "Delete a fosil by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the fosil to delete",
                        "name": "fosilId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.FosilResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.FosilResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.FosilResponse"
                        }
                    }
                }
            }
        },
        "/fosils": {
            "get": {
                "description": "Retrieve all fosils from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fosil"
                ],
                "summary": "Get all fosils",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.FosilResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.FosilResponse"
                        }
                    }
                }
            }
        },
        "/fosils/export": {
            "get": {
                "description": "Get data fosil from MongoDB and export to excel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fosil"
                ],
                "summary": "Export data fosil to excel",
                "responses": {
                    "200": {
                        "description": "Data fosil exported to excel successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.FosilResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responses.FosilResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}